@@ -7,16 +7,17 @@ issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+
 version: 5.1.1
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
-  flutter: ">=2.8.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
   plugin_platform_interface: ^2.1.0
+  async: ^2.11.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   pedantic: ^1.10.0
-  pigeon: 0.1.21
+  pigeon: ^10.0.0
