@@ -6,11 +6,11 @@ import 'dart:ui';
 
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:video_player_platform_interface/messages.dart';
+import 'package:video_player_platform_interface/messages.g.dart';
 import 'package:video_player_platform_interface/method_channel_video_player.dart';
 import 'package:video_player_platform_interface/video_player_platform_interface.dart';
 
-import 'test.dart';
+import 'test.g.dart';
 
 class _ApiLogger implements TestHostVideoPlayerApi {
   final List<String> log = <String>[];
@@ -26,7 +26,7 @@ class _ApiLogger implements TestHostVideoPlayerApi {
   TextureMessage create(CreateMessage arg) {
     log.add('create');
     createMessage = arg;
-    return TextureMessage()..textureId = 3;
+    return TextureMessage(textureId: 3);
   }
 
   @override
@@ -62,7 +62,7 @@ class _ApiLogger implements TestHostVideoPlayerApi {
   PositionMessage position(TextureMessage arg) {
     log.add('position');
     textureMessage = arg;
-    return PositionMessage()..position = 234;
+    return PositionMessage(position: 234, textureId: arg.textureId);
   }
 
   @override
@@ -87,6 +87,11 @@ class _ApiLogger implements TestHostVideoPlayerApi {
   void setPlaybackSpeed(PlaybackSpeedMessage arg) {
     log.add('setPlaybackSpeed');
     playbackSpeedMessage = arg;
+  }
+
+  @override
+  AbsolutePositionMessage absolutePosition(TextureMessage arg) {
+    throw UnimplementedError();
   }
 }
 
@@ -233,6 +238,13 @@ void main() {
       expect(log.log.last, 'position');
       expect(log.textureMessage?.textureId, 1);
       expect(position, const Duration(milliseconds: 234));
+    });
+
+    test('getAbsolutePosition', () async {
+      final int? absolutePosition = await player.getAbsolutePosition(1);
+      expect(log.log.last, 'absolutePosition');
+      expect(log.textureMessage?.textureId, 1);
+      expect(absolutePosition, const Duration(milliseconds: 234));
     });
 
     test('videoEventsFor', () async {
